TV\Radio Schedules: Shift solution Tab 9 added except output functionality

+++ Business/Business.Solutions/Common/Entities/NonPersistent/ImageClipartObject.cs
@@ -10,16 +10,22 @@ namespace Asa.Business.Solutions.Common.Entities.NonPersistent
 		public override ClipartObjectType Type => ClipartObjectType.Image;
 
 		public Image Image { get; set; }
+		public string FilePath { get; set; }
 
 		public static ImageClipartObject FromFile(string filePath)
 		{
-			var tempFile = Path.GetTempFileName();
-			File.Copy(filePath, tempFile, true);
-			return new ImageClipartObject
+			if (File.Exists(filePath))
 			{
-				Name = Path.GetFileNameWithoutExtension(filePath),
-				Image = Image.FromFile(filePath)
-			};
+				var tempFile = Path.GetTempFileName();
+				File.Copy(filePath, tempFile, true);
+				return new ImageClipartObject
+				{
+					Name = Path.GetFileNameWithoutExtension(filePath),
+					FilePath = filePath,
+					Image = Image.FromFile(filePath)
+				};
+			}
+			return null;
 		}
 
 		public static ImageClipartObject FromImageSource(ImageSource imageSource)
@@ -27,6 +33,7 @@ namespace Asa.Business.Solutions.Common.Entities.NonPersistent
 			return new ImageClipartObject
 			{
 				Name = imageSource.Name,
+				FilePath = imageSource.FileName,
 				Image = imageSource.OriginalImage ?? imageSource.BigImage
 			};
 		}
